fix: refactor acording to review
