+++ package.json
@@ -51,7 +51,6 @@
     "@govtechsg/tradetrust-utils": "^1.0.1",
     "@reduxjs/toolkit": "^1.6.1",
     "@types/gtag.js": "0.0.8",
-    "babel-plugin-transform-require-context": "^0.1.1",
     "cross-env": "^7.0.3",
     "date-fns": "^2.22.1",
     "date-fns-tz": "^1.1.4",
@@ -108,6 +107,7 @@
     "babel-jest": "^27.0.2",
     "babel-loader": "^8.2.2",
     "babel-plugin-macros": "^3.1.0",
+    "babel-plugin-transform-require-context": "^0.1.1",
     "brotli-webpack-plugin": "^1.1.0",
     "compression-webpack-plugin": "^6.1.1",
     "concurrently": "^6.2.0",
