+++ src/pages/event/index.tsx
@@ -1,0 +1,6 @@
+import { getSortedByDateDesc, getCmsContentWithSlug } from "../../utils";
+
+export * from "./event";
+
+const rawEvents = getCmsContentWithSlug(require.context("../../../cms/event/", false, /\.md$/));
+export const events = getSortedByDateDesc(rawEvents);
